added non null check for TankerService when terminate action is performed changed the style of tanker terminate button so that it shows a red background and white front, closes #9

+++ src/main/java/moe/ofs/addon/aarservice/Dispatcher.java
@@ -298,6 +298,10 @@ public class Dispatcher {
 
     // manually destroy tanker object and data
     public void terminateDispatch(TankerService service) {
+
+        // FIXME
+        if(service.getUnit() == null) return;
+
         // destroy tanker immediately
         new ServerActionRequest(String.format("Unit.destroy({ id_ = %d })", service.getUnit().getId())).send();
